remove config
