+++ server/src/main/java/org/apache/iotdb/db/conf/IoTDBDescriptor.java
@@ -1034,6 +1034,8 @@ public class IoTDBDescriptor {
     conf.setQuotaEnable(
         Boolean.parseBoolean(
             properties.getProperty("quota_enable", String.valueOf(conf.isQuotaEnable()))));
+
+    conf.setRateLimiterType(properties.getProperty("rate_limiter_type", conf.getRateLimiterType()));
   }
 
   private void loadAuthorCache(Properties properties) {
